requirements/common.txt
Django==1.8.1
django-imagekit==3.2.6
